requirements.txt
pygame>=2.5.0
pillow>=10.0.0
